R ess2 (#75)
ess is a new container with R built from source (with X11 support). The container has emacs and ess for a lightweight R IDE which requires less work to get started compared to the Rstudio server containers. For now, this container comes with one tag (other than dev), R version 4.1.0 (and packages frozen to the same date on CRAN as the rocker 4.1.0 packages by using the RSPM with a specific date as the CRAN repo). Note that in order to actually use X11 in the container (e.g. to plot at the terminal and get the window to come through to your local machine), the server the container is running on also needs X11. On HPC, I recommend finding an R module or other module that has X11 included (not all R modules are build with X11, but you might be able to find one that does...). You can tell if X11 is included and working in an R installation by running capabilities(). Loading that R module on the host will also load X11 on the host. Then spin up the container and you should be able to make plots from the terminal in the container.